+++ Dockerfile
@@ -1,6 +1,6 @@
 FROM openjdk:11-jre
 
-ADD target/signservice-ca-1.0.0-SNAPSHOT.jar /app.jar
+ADD target/signservice-ca-1.0.0.jar /app.jar
 ENTRYPOINT ["java","-jar","/app.jar"]
 
 # Main web port
