+++ pom.xml
@@ -19,7 +19,7 @@
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-    <ca.service.base.version>1.3.5-SNAPSHOT</ca.service.base.version>
+    <ca.service.base.version>1.4.0</ca.service.base.version>
     <sivaltrust.commons.version>1.0.2</sivaltrust.commons.version>
     <java.version>11</java.version>
   </properties>
